+++ packages/odata-common/src/expandable.ts
@@ -7,7 +7,7 @@ import { EntityApi } from './entity-api';
 
 /**
  * Represents everything that can be used in an `.expand` statement. Only relevant for OData v4 requests.
- * @typeparam EntityT - Type of the entity to be selected on
+ * @typeParam EntityT - Type of the entity to be selected on
  */
 export type Expandable<
   EntityT extends EntityBase,