+++ packages/odata-common/src/filter/filter-function.ts
@@ -4,27 +4,18 @@ import { NumberFilterFunction } from './number-filter-function';
 import { StringFilterFunction } from './string-filter-function';
 import { FilterFunctionParameterType } from './filter-function-base';
 
-/**
- * @internal
- */
 export function filterFunction<EntityT extends EntityBase>(
   functionName: string,
   returnType: 'boolean',
   ...parameters: FilterFunctionParameterType<EntityT>[]
 ): BooleanFilterFunction<EntityT>;
 
-/**
- * @internal
- */
 export function filterFunction<EntityT extends EntityBase>(
   functionName: string,
   returnType: 'int' | 'double' | 'decimal',
   ...parameters: FilterFunctionParameterType<EntityT>[]
 ): NumberFilterFunction<EntityT>;
 
-/**
- * @internal
- */
 export function filterFunction<EntityT extends EntityBase>(
   functionName: string,
   returnType: 'string',
@@ -33,11 +24,10 @@ export function filterFunction<EntityT extends EntityBase>(
 
 /**
  * Build a custom filter function.
- * @param functionName - the name of the function, e.g., `substring`
- * @param returnType - the return type of the filter function
- * @param parameters - the parameter(s) used in the function
- * @returns An instance of filter function suited for the given return type
- * @internal
+ * @param functionName - The name of the function, e.g., `substring`.
+ * @param returnType - The return type of the filter function.
+ * @param parameters - The parameter(s) used in the function.
+ * @returns An instance of filter function suited for the given return type.
  */
 export function filterFunction<EntityT extends EntityBase>(
   functionName: string,