+++ packages/odata-common/src/request-builder/update-request-builder-base.ts
@@ -16,7 +16,7 @@ import { MethodRequestBuilder } from './request-builder-base';
 
 /**
  * Abstract class to create OData query to update an entity containing methods shared for OData v2 and v4.
- * @typeparam EntityT - Type of the entity to be updated
+ * @typeParam EntityT - Type of the entity to be updated
  */
 export abstract class UpdateRequestBuilderBase<
     EntityT extends EntityBase,