+++ packages/odata-v2/src/entity.ts
@@ -1,8 +1,8 @@
-import { EntityBase } from '@sap-cloud-sdk/odata-common/internal';
+import { EntityBase } from '@sap-cloud-sdk/odata-common';
 
 /**
  * Super class for all representations of OData v2 entity types.
  */
 export class Entity extends EntityBase {
-  readonly _oDataVersion: 'v2' = 'v2';
+  readonly _oDataVersion: 'v2' = 'v2' as const;
 }