+++ packages/odata-v4/src/entity.ts
@@ -1,8 +1,8 @@
-import { EntityBase } from '@sap-cloud-sdk/odata-common/internal';
+import { EntityBase } from '@sap-cloud-sdk/odata-common';
 
 /**
  * Super class for all representations of OData v4 entity types.
  */
 export class Entity extends EntityBase {
-  readonly _oDataVersion: 'v4' = 'v4';
+  readonly _oDataVersion: 'v4' = 'v4' as const;
 }