+++ test-packages/test-services-odata-v2/test-service/TestLvl2NestedComplexType.ts
@@ -32,7 +32,7 @@ export interface TestLvl2NestedComplexType<
 
 /**
  * TestLvl2NestedComplexTypeField
- * @typeparam EntityT - Type of the entity the complex type field belongs to.
+ * @typeParam EntityT - Type of the entity the complex type field belongs to.
  */
 export class TestLvl2NestedComplexTypeField<
   EntityT extends Entity,