+++ packages/util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-cloud-sdk/util
 
+## 2.7.1
+
+## 2.7.0
+
+## 2.6.0
+
 ## 2.5.0
 
 ### Minor Changes